address review comment
