+++ api/client/alpn_conn_upgrade_test.go
@@ -21,13 +21,10 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"errors"
-	"io"
 	"net"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
-	"sync"
-	"sync/atomic"
 	"testing"
 	"time"
 
@@ -35,6 +32,7 @@ import (
 
 	"github.com/gravitational/teleport/api/constants"
 	"github.com/gravitational/teleport/api/fixtures"
+	"github.com/gravitational/teleport/api/testhelpers"
 	"github.com/gravitational/teleport/api/utils/pingconn"
 )
 
@@ -73,7 +71,7 @@ func TestIsALPNConnUpgradeRequired(t *testing.T) {
 		},
 	}
 
-	forwardProxy := mustStartForwardProxy(t)
+	forwardProxy, forwardProxyURL := mustStartForwardProxy(t)
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
@@ -83,9 +81,9 @@ func TestIsALPNConnUpgradeRequired(t *testing.T) {
 			})
 
 			t.Run("with ProxyURL", func(t *testing.T) {
-				countBeforeTest := forwardProxy.numProxiedConnections.Load()
-				require.Equal(t, test.expectedResult, IsALPNConnUpgradeRequired(server.Addr().String(), test.insecure, forwardProxy.dialOption()))
-				require.Equal(t, countBeforeTest+1, forwardProxy.numProxiedConnections.Load())
+				countBeforeTest := forwardProxy.Count()
+				require.Equal(t, test.expectedResult, IsALPNConnUpgradeRequired(server.Addr().String(), test.insecure, withProxyURL(forwardProxyURL)))
+				require.Equal(t, countBeforeTest+1, forwardProxy.Count())
 			})
 		})
 	}
@@ -189,10 +187,10 @@ func TestALPNConnUpgradeDialer(t *testing.T) {
 			})
 
 			t.Run("with ProxyURL", func(t *testing.T) {
-				forwardProxy := mustStartForwardProxy(t)
-				countBeforeTest := forwardProxy.numProxiedConnections.Load()
+				forwardProxy, forwardProxyURL := mustStartForwardProxy(t)
+				countBeforeTest := forwardProxy.Count()
 
-				proxyURLDialer := newProxyURLDialer(forwardProxy.url, directDialer)
+				proxyURLDialer := newProxyURLDialer(forwardProxyURL, directDialer)
 				dialer := newALPNConnUpgradeDialer(proxyURLDialer, tlsConfig, test.withPing)
 				conn, err := dialer.DialContext(ctx, "tcp", addr.Host)
 				if test.wantError {
@@ -203,7 +201,7 @@ func TestALPNConnUpgradeDialer(t *testing.T) {
 				defer conn.Close()
 
 				mustReadConnData(t, conn, "hello")
-				require.Equal(t, countBeforeTest+1, forwardProxy.numProxiedConnections.Load())
+				require.Equal(t, countBeforeTest+1, forwardProxy.Count())
 			})
 		})
 	}
@@ -313,60 +311,7 @@ func mockConnUpgradeHandler(t *testing.T, upgradeType string, write []byte) http
 	})
 }
 
-// forwardProxy is a simple implementation of a HTTPS_PROXY.
-type forwardProxy struct {
-	net.Listener
-	url                   *url.URL
-	numProxiedConnections atomic.Uint32
-}
-
-func (p *forwardProxy) dialOption() DialOption {
-	return func(cfg *dialConfig) {
-		cfg.proxyURLFunc = func(_ string) *url.URL {
-			return p.url
-		}
-	}
-}
-
-func (p *forwardProxy) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
-	if req.Method != http.MethodConnect {
-		rw.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	// Prepare serverConn.
-	serverConn, err := net.Dial("tcp", req.Host)
-	if err != nil {
-		rw.WriteHeader(http.StatusServiceUnavailable)
-		return
-	}
-	defer serverConn.Close()
-
-	// Let client know it's ready.
-	rw.WriteHeader(http.StatusOK)
-
-	// Prepare clientConn.
-	hijacker, _ := rw.(http.Hijacker)
-	clientConn, _, _ := hijacker.Hijack()
-	defer clientConn.Close()
-
-	// Start proxy conn.
-	p.numProxiedConnections.Add(1)
-
-	wg := sync.WaitGroup{}
-	wg.Add(2)
-	proxyConn := func(dst, src io.ReadWriteCloser) {
-		defer dst.Close()
-		defer src.Close()
-		io.Copy(dst, src)
-		wg.Done()
-	}
-	go proxyConn(serverConn, clientConn)
-	go proxyConn(clientConn, serverConn)
-	wg.Wait()
-}
-
-func mustStartForwardProxy(t *testing.T) *forwardProxy {
+func mustStartForwardProxy(t *testing.T) (*testhelpers.ProxyHandler, *url.URL) {
 	t.Helper()
 
 	listener, err := net.Listen("tcp", "localhost:0")
@@ -378,10 +323,7 @@ func mustStartForwardProxy(t *testing.T) *forwardProxy {
 	url, err := url.Parse("http://" + listener.Addr().String())
 	require.NoError(t, err)
 
-	forwardProxy := &forwardProxy{
-		Listener: listener,
-		url:      url,
-	}
-	go http.Serve(listener, forwardProxy)
-	return forwardProxy
+	handler := &testhelpers.ProxyHandler{}
+	go http.Serve(listener, handler)
+	return handler, url
 }
